Separate responsibility and decrease dependency between classes
