+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
